ref(*)!: Reorganizes the blobstore interface
This is a relatively simple PR that mostly restructures things. This
removes the container interface because there wasn't much you could do
with it by itself, it always had to be used in conjunction with the
blobstore interface. I also got rid of all of the type aliases for
strings to make the types interface cleaner, as well as follow what we've
been doing in the other wasi cloud interfaces with avoiding type aliases.

Additionally, I updated the timestamps to use the wasi-clock timestamps
instead so we were using a concrete type. That was the only major change
here, but it does solve an outstanding issue we have.

Closes #7

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>